move cursor to module
