+++ lib/paginator/config.ex
@@ -1,8 +1,6 @@
 defmodule Paginator.Config do
   @moduledoc false
 
-  alias Paginator.Cursor
-
   @type t :: %__MODULE__{}
 
   defstruct [
@@ -11,6 +9,7 @@ defmodule Paginator.Config do
     :before,
     :before_values,
     :cursor_fields,
+    :cursor_module,
     :include_total_count,
     :limit,
     :maximum_limit,
@@ -22,19 +21,25 @@ defmodule Paginator.Config do
   @minimum_limit 1
   @maximum_limit 500
   @default_total_count_limit 10_000
+  @default_cursor_module Paginator.Cursors.UnencryptedCursor
 
   def new(opts \\ []) do
     %__MODULE__{
       after: opts[:after],
-      after_values: Cursor.decode(opts[:after]),
+      after_values: cursor_module(opts).decode(opts[:after]),
       before: opts[:before],
-      before_values: Cursor.decode(opts[:before]),
+      before_values: cursor_module(opts).decode(opts[:before]),
       cursor_fields: opts[:cursor_fields],
+      cursor_module: cursor_module(opts),
       include_total_count: opts[:include_total_count] || false,
       limit: limit(opts),
       sort_direction: opts[:sort_direction] || :asc,
       total_count_limit: opts[:total_count_limit] || @default_total_count_limit
     }
+  end
+
+  defp cursor_module(opts) do
+    opts[:cursor_module] || @default_cursor_module
   end
 
   defp limit(opts) do
