+++ lib/paginator/cursor.ex
@@ -1,21 +1,4 @@
 defmodule Paginator.Cursor do
-  @moduledoc false
-
-  def decode(nil), do: nil
-
-  def decode(encoded_cursor) do
-    encoded_cursor
-    |> Base.url_decode64!()
-    |> :erlang.binary_to_term()
-  end
-
-  def encode(values) when is_list(values) do
-    values
-    |> :erlang.term_to_binary()
-    |> Base.url_encode64()
-  end
-
-  def encode(value) do
-    encode([value])
-  end
+  @callback decode(String.t()) :: {:ok, term :: term} | {:error, term :: term}
+  @callback encode(arg :: term) :: {:ok, encoded :: String.t()} | {:error, String.t()}
 end
